Move BLAS/LAPACK to Platforms/CPU.
