+++ src/AFQMC/Estimators/SlaterDetOperations.h
@@ -6,7 +6,6 @@
 #include "AFQMC/config.h"
 #include <Message/MPIObjectBase.h>
 #include "Numerics/DeterminantOperators.h"
-#include "Numerics/Blasf.h"
 #include "Numerics/MatrixOperators.h"
 #include "Message/Communicate.h"
 //#include "Message/CommOperators.h"
