+++ src/Particle/LongRange/StructFact.cpp
@@ -17,7 +17,7 @@
 #include <config/stdlib/math.hpp>
 #include <Numerics/e2iphi.h>
 #include <CPU/SIMD/vmath.hpp>
-#include <Numerics/OhmmsBlas.h>
+#include <CPU/BLAS.hpp>
 #include <qmc_common.h>
 
 namespace qmcplusplus
