+++ src/Platforms/CPU/CMakeLists.txt
@@ -11,9 +11,10 @@
 
 
 SET(CPU_SRCS
+  BlasThreadingEnv.cpp
 )
 
-ADD_LIBRARY(platform_cpu INTERFACE)
+ADD_LIBRARY(platform_cpu ${CPU_SRCS})
 TARGET_INCLUDE_DIRECTORIES(platform_cpu INTERFACE "${CMAKE_CURRENT_SOURCE_DIR}")
 
 IF (BUILD_UNIT_TESTS)
