+++ src/QMCDrivers/WFOpt/QMCLinearOptimize.cpp
@@ -25,7 +25,7 @@
 //#include "QMCDrivers/VMC/VMCSingle.h"
 //#include "QMCDrivers/QMCCostFunctionSingle.h"
 #include "QMCHamiltonians/HamiltonianPool.h"
-#include "Numerics/Blasf.h"
+#include "CPU/Blasf.h"
 #include "Numerics/MatrixOperators.h"
 #include <cassert>
 #if defined(QMC_CUDA)
