+++ src/QMCHamiltonians/MomentumEstimator.cpp
@@ -16,7 +16,7 @@
 #include <QMCHamiltonians/MomentumEstimator.h>
 #include <QMCWaveFunctions/TrialWaveFunction.h>
 #include <Numerics/e2iphi.h>
-#include <Numerics/OhmmsBlas.h>
+#include <CPU/BLAS.hpp>
 #include <OhmmsData/AttributeSet.h>
 #include <Utilities/SimpleParser.h>
 #include <Particle/DistanceTableData.h>
