+++ src/QMCWaveFunctions/Fermion/DiracDeterminantBatched.cpp
@@ -12,7 +12,7 @@
 
 #include "QMCWaveFunctions/Fermion/DiracDeterminantBatched.h"
 #include "Numerics/DeterminantOperators.h"
-#include "Numerics/OhmmsBlas.h"
+#include "CPU/BLAS.hpp"
 #include "Numerics/MatrixOperators.h"
 #include "CPU/SIMD/simd.hpp"
