+++ src/QMCWaveFunctions/Fermion/DiracMatrix.h
@@ -12,9 +12,9 @@
 #ifndef QMCPLUSPLUS_DIRAC_MATRIX_H
 #define QMCPLUSPLUS_DIRAC_MATRIX_H
 
-#include "Numerics/Blasf.h"
+#include <CPU/Blasf.h>
+#include <CPU/BlasThreadingEnv.h>
 #include <OhmmsPETE/OhmmsMatrix.h>
-#include "Numerics/BlasThreadingEnv.h"
 #include <type_traits/scalar_traits.h>
 #include "CPU/SIMD/simd.hpp"
